backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
sqlalchemy>=2.0.41,<2.1
python-dotenv==1.0.1
alembic==1.13.2
psycopg[binary]==3.2.13
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.2.0
httpx==0.27.2
pytest==8.3.3
python-multipart==0.0.9
